Expose function to define new signals.
