+++ CMakeLists.txt
@@ -44,6 +44,7 @@ SET(${PROJECT_NAME}_HEADERS
   include/${CUSTOM_HEADER_DIR}/interpreter.hh
   include/${CUSTOM_HEADER_DIR}/module.hh
   include/${CUSTOM_HEADER_DIR}/python-compat.hh
+  include/${CUSTOM_HEADER_DIR}/signal.hh
   include/${CUSTOM_HEADER_DIR}/signal-wrapper.hh
   )
